+++ libs/storage/src/providers/rustfs-storage.provider.ts
@@ -1,27 +1,30 @@
 import { BaseS3StorageService } from './base-s3-storage.provider';
 import { Injectable } from '@nestjs/common';
+import { S3 } from 'aws-sdk';
 
 @Injectable()
 export class RustFsStorageService extends BaseS3StorageService {
-  constructor() {
-    const endpoint = process.env.RUSTFS_ENDPOINT || 'http://localhost:9000';
+  private readonly endpoint = process.env.RUSTFS_ENDPOINT || 'http://localhost:9000';
+  private readonly baseConfig = {
+    accessKeyId: process.env.RUSTFS_ACCESS_KEY_ID,
+    secretAccessKey: process.env.RUSTFS_SECRET_ACCESS_KEY,
+    endpoint: this.endpoint,
+    s3ForcePathStyle: true
+  };
 
-    const baseConfig = {
-      accessKeyId: process.env.RUSTFS_ACCESS_KEY_ID,
-      secretAccessKey: process.env.RUSTFS_SECRET_ACCESS_KEY,
-      endpoint,
-      s3ForcePathStyle: true
-    };
+  protected async getS3Client(): Promise<S3> {
+    return new S3({ ...this.baseConfig, region: process.env.AWS_REGION });
+  }
 
-    const s3Config = { ...baseConfig, region: process.env.AWS_REGION };
-    const s4Config = { ...baseConfig, region: process.env.AWS_PUBLIC_REGION };
-    const storeObjectConfig = { ...baseConfig, region: process.env.AWS_S3_STOREOBJECT_REGION };
+  protected async getPublicS3Client(): Promise<S3> {
+    return new S3({ ...this.baseConfig, region: process.env.AWS_PUBLIC_REGION });
+  }
 
-    super(s3Config, s4Config, storeObjectConfig);
+  protected async getStoreObjectS3Client(): Promise<S3> {
+    return new S3({ ...this.baseConfig, region: process.env.AWS_S3_STOREOBJECT_REGION });
   }
 
   getPublicUrl(bucketName: string, fileKey: string): string {
-    const endpoint = process.env.RUSTFS_ENDPOINT || 'http://localhost:9000';
-    return `${endpoint}/${bucketName}/${fileKey}`;
+    return `${this.endpoint}/${bucketName}/${fileKey}`;
   }
 }
